TEST-26 Updated with review comments
