+++ dominion-pax-exam-driver/src/main/java/org/codice/dominion/pax/exam/DominionPaxExam.java
@@ -28,6 +28,7 @@ import org.ops4j.pax.exam.Constants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/** JUnit runner for the Dominion PaxExam driver. */
 public class DominionPaxExam extends Runner implements Filterable, Sortable {
   private static final Logger LOGGER = LoggerFactory.getLogger(DominionPaxExam.class);
 
@@ -71,7 +72,8 @@ public class DominionPaxExam extends Runner implements Filterable, Sortable {
     return "DominionPaxExam[" + testClass.getName() + ", " + delegate + "]";
   }
 
-  @SuppressWarnings("squid:CommentedOutCodeLine" /* left over code from PaxExam */)
+  @SuppressWarnings(
+      "squid:CommentedOutCodeLine" /* left over code from PaxExam to be removed when we are guaranteed we won't support CDI */)
   private static ParentRunner createDelegate(Class<?> testClass) throws InitializationError {
     // force us into a per-suite strategy
     System.setProperty(
