+++ dominion-pax-exam-driver/src/main/java/org/codice/dominion/pax/exam/internal/DominionProbeRunner.java
@@ -180,7 +180,7 @@ public class DominionProbeRunner extends BlockJUnit4ClassRunner {
     Statement statement = childrenInvoker(notifier);
 
     if (!filteredChildren().allMatch(this::isIgnored)) {
-      LOGGER.debug("{}::1lassBlock() - adding @BeforeExam and @AfterExam statements", this);
+      LOGGER.debug("{}::classBlock() - adding @BeforeExam and @AfterExam statements", this);
       statement = withBeforeClasses(statement);
       statement = withAfterClasses(statement);
     } else {
